refactor(Alert): create typscript version of component and now use react hook
